Add support for openbmc push notify, CSM alert config
This build add support for openbmc push notifications using websockets.
The websocket will be setup automatically when the openBMC interface
is configured in the configuration file. On startup, the crassd service
will open a websocket in a thread for each monitored server using them.
If the websocket is closed, the crassd service will attempt to re-open
the websocket on the next polling interval. If unable to open the
websocket, the crassd will issue an error that it is unable to
communicate with the bmc. This alert will only be sent once.

The CSM plugin also has changes for how alerts are forwarded to it.
First, the plugin will allow configuration of what is forwarded via a
policy table. This can be found in the plugin's directory. The second
change is to allow CSM to populate messages with slot/component numbers
into a placeholder. This support requires 1.04 or greater of the
openbmctool.

Signed-off-by: Justin Thaler thalerj@us.ibm.com