Update ESS status on crassd initialization
This update provides an initialization function so ess does not
continually report a checking status. At startup, if a checking status
is shown by ESS, a healthy alert is sent for each node, initializing
the status for each node to 'HEALTHY'. This should provide better user
experience.

An update was made to the config file to add a varaible containing the
notifyList so the ess plugin can use it.

An update was made to ibm-crassd to load the nodes before initializing
plugins.

Signed-Off-By: Justin Thaler thalerj@us.ibm.com