Fixes ess plugin and ESS install env
This provides fixes for initializing the plugin without throwing an
exception. It also fixes notifying ess by fixing a coded dictionary key,
such that ess is now used instead of mmhealth.

the main file for the application has been updated so that the notify
list dictionary is tied to the config module, and so the notification
listener file is only loaded when needed. This means not loading it when
there are no openbmc systems used.

Signed-off-by: Justin Thaler thalerj@us.ibm.com