Update event modifier to be easier to read
Updates the openbmc event dictionary modifier that provides common keys
between the openbmc events and the ipmi sourced events to be easier to
read. It provides an easier way to add additional keys which are defined
in a dictionary and then loops through the dictionary to provide the
right keys for the event parser.

Signed-off-by: Justin Thaler thalerj@us.ibm.com