Generalize rpmbuild files
A generic spec is now used instead of having an arch dependent one.
The generated rpm should now be noarch.
Moved the docker and compose files to an example directory, and the
Docker build will now need an rpm instead of relying on make/install.
Added a build script for the rpm generation.